Fix the b* prototypes
Use float for window start and end.

+++ src/operators/balw.hpp
@@ -10,8 +10,8 @@ extern "C" {
 void
 stl_balw(const thrust::device_ptr<sigpt_t> &in,
          const int nin,
-         const int s,
-         const int t,
+         const float s,
+         const float t,
          thrust::device_ptr<sigpt_t> *out,
          int *nout);
